refactor: improve entry edit sheet UX and fix data integrity
- Extract EntryEditSheet to its own file
- Fix bug where editing summary overwrote original content
- Add empty-summary guard in save closure
- Add "1 = highest" hint to priority picker
- Use displayName instead of rawValue for category labels
- Show current category first in picker, remove question/thought
- Remove stale onEdit parameter from DevScreen
- Fix Makefile entitlements validation to check for build variable
